added more text instructions
will be changing the instructions just wanted to put a place holder  and the game over screen for some reason doesn't work; it worked with the previous code before the change to level 1.  I tried to see where the problem was, but I couldn't find anything.  Hopefully we'll get to the bottom of this